add `--disable_distributed_sampler_fallback` option
Multipack may fail either because the dataset is too small for the
number of GPUs OR because the packing was unsuccessful.
Currently, the loop will fall back to a naive DistributedSampler.
This alters the default behavior, allowing the user to cause the loop to
"fail fast" when the Multipack sampler fails.

Signed-off-by: James Kunstle <jkunstle@redhat.com>